Always set Representing
For Outlook, it seems that it is necessary to set the Representing
properties all the time, regardless of whether the message is being
sent on behalf of someone else.  Outlook seems to be expecting these
fields to be present.  If they are not, the From field in the folder
messages lists will be blank (Note, if you pop open the message, the
from fields will be filled out).