Remove Setting SentOn = DateTime.UtcNow
It is better to leave this not set rather then setting the time to now,
as it is too assuming about the time of the message.  I came across
this processing emails from 12 years ago and they were showing up as
being created today.  So, setting the time to now, would be incorrect
and just confusing.  Better to leave it not set, as it will indicate
incomplete information.